requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
streamlit
plotly
scikit-learn
statsmodels
joblib
python-dotenv
reportlab